+++ app/XDM.Android/core-model/src/main/kotlin/com/mikeyphw/xdm/android/model/DownloadIntake.kt
@@ -10,6 +10,7 @@ enum class DownloadIntakeOrigin {
     ExternalShare,
     ExternalView,
     ExternalDownloadManager,
+    BrowserExtension,
     Automation,
     Clipboard,
     ManualEntry,